move http(s) server creation to own file
